Add files via upload
Insights for Breast Cancer Diagnosis:
By comparing all the four values of k i.e. 21,11,33 and 9. We can analyze that even though the accuracy of the k-NN algorithm is highest i.e. 98% when the k-value is 21 as well as 11, it has a type 2 error in the case when the k-value is 21 while it has both type 1 and type 2 error when the k-value is 11. When the k value is increased beyond 21 to 33 the accuracy decreases to 97% and even if the k value is decreased to 9 the accuracy drops down to 97%. Thus, an optimal k value for this case would 21 as it is giving the highest accuracy and doesn’t have type 1 error. From the medical point of view, the value of 21 for k would diagnose the results more accurately and classify the cancer cells appropriately.